Stand-up first support for DynamoDB.
This patch adds initial support for DynamoDB as a backend. So far,
I've only tested it with local ScyllaDB instances over the Alternator
interface.

I've also scrubbed the codebase of TODOs & lints, and added a new
binary, `indielinks-ddb` that will create the DDB schema and
charge the database with initial data.

Hitting `/dev/` on the local host continues to fetch user `sp1ff`.