Add a webfinger handler.
This patch will add support for serving requests at
`/.well-known/webfinger`, along with supporting code for the `acct`
URI scheme and some initial HTTP plumbing for indielinks.

The webfinger implementation is minimal; it is, as I recall, the
minimum necessary to allow Mastodon to "see", but not follow
indielink users.