Implement `posts/get`, `posts/delete` and `tags/get`.
This (rather large) patch contains a first implementation of the
`posts/get` and `posts/delete` handlers. I'm unhappy with much
of the implementation of the storage backends, but I've wanted
to have a working integration test suite before I re-factor.

The patch includes a start on the Indielinks metrics framework;
I've used Tolnay's `inventory` crate to support registering
metrics near their use and macros to reduce the boilerplate
involved looking-up the OTel instruments and calling their
various methods.

It also includes a start on integration tests for these two
endpoints, building on the work in the last patch. I've laid-out
the basic structure of the integration tests, the configuration
facility, & so on, and implemented a first "smoke" test for
the del.icio.us API.