Implement the `/users/signup` endpoint.
This is a very large patch that not only implements the HTTP
`/users/signup` endpoint, but the infrastructure behind it for
storing passwords. indielinks now salts and peppers passwords
before using Argon2id to hash them, with provisions for rotating
the pepper.

I also wound-up cleaning-up the implementation of my various
refinement types in the `entities` module; I have to wonder
if this shouldn't be wrapped-up in a macro or library, somehow.

Simple integration test implemented, as well.