Implement the ActivityPub actor endpoint.
This patch adds the `/users/{username}` handler, which is the
top-level ActivityPub "activity" endpoint for each user. The
per-user inbox, outbox, and so forth will found beneath this.
This patch supports returning either the JRD (if the Accept
header is set to "application/activity+json") or a trivial
HTML page if it's set to "text/html".