Accept ActivityPub follows.
This (rather large) patch implements a handler for the
`/users/{username}/inbox` endpoint that handles follow requests
(including sending the accept message back). HTTP request signature
verification & creation are also implemented.

It also adds a script to the `infra` folder, `stunnel-up`. This
will create a TLS endpoint on the localhost at port 443 and
forward the traffic to the local indielink's public endpoint.
It uses a self-signed certificate (also part of this commit).

I've tested against a local Mastodon instance and stood-up a minimal
integration test.