Standarize the way indielinks implements refined types in Rust and normalizes the handling of the indielinks origin.
This commit does three things. Firstly, it adds an `id` column
to the `posts` table.

Secondly, it updates all refined types to be implemented "by hand",
according to a checklist documented in `_doc.rs`.

Lastly, it eliminates the assorted uses of "protocol", "domain",
"host", and so forth and introduces a singled type `Origin` to
represent the indielinks origin. All the ActivityPub logic that forms
indielinks URLs now works in terms of `Origin`.