Serve Posts over ActivityPub & add middleware to the indielinks http clients
Serve a `Note` corresponding to a Post via ActivityPub. This commit
also sets-up a standard HTTP client for indielinks to use. It uses the
`reqwest-middleware` crate to do a few things for that client:

    - automatic retries
    - emit metrics
    - optionally sign outgoing requests

This resulted in a great deal of clean-up in the request signing
logic, as well.

Integration testing is still needed, however.