Allow indielinks users to send Follow requests.
This patch adds the endpoint `/api/v1/users/follow`, which will send
an ActivityPub Follow request to a given user. The endpoint will only
write a `SendFollow` task to the data store. All the work of writing-down
the follow, and sending the AP activity will be handled in the background
task.

Still to be implemented: handling the resulting Accept message.