Initial commit of `indielinks-cache`.
This patch adds a new library crate to the workspace: `indielinks-cache`.
It uses the `openraft` crate to stand-up a very basic Raft cluster that
can be extended to meet indielinks' needs. It includes a simple implementation
of `main()` that stands-up an HTTP server supporting the cluster.

At this commit, one can configure the cluster (i.e. make the nodes
aware of one another & cause them to elect a leader). See the
`exercise-cluster` script for how to do that.