Add the "application" interface to `indielinks-cache`.
The cache now maintains the hash ring, albeit in a trivial way:
it's initialized once when the Raft cluster is initialized, and
re-initialized every time membership changes. Clients of the
sample app can add (but not remove) node mappings to the ring.

This patch includes a trivial test in the `exercise-cluster`
shell script.