Serve static assets from indielinks.
This patch introduces a new route, `/fe` (for front-end) that
serves static assets expected to be found in an `assets` subdirectory
of the process' working directory. The intent is to use this in
the future to serve the frontend directly from indielinks, but
this will need a formal "release" process.