Revamp the indielinks metrics implementation.
After realizing that the actual OTel instruments are only used
at the site of measurement, and that I really just wanted to
check for name collisions, I re-wrote my metric implementation.

The last commit implemented the new logic. This commit moves
the entire code base over to that new implementation.