Accurately check the "private" & "unread" boxes when editing posts. Add an "unread" parameter to `/posts/all`.
This patch fixes a bug on the Home page where, when editing a post,
the "private" and "unread" boxes were always unchecked, no matter
what the values of those attributes were. This patch fixes that.

It also adds an indielinks-specific extension to the set of
parameters on which one can filter when calling `/posts/all`--
"unread". If this is set to true, only unread posts will be
returned.