Add an Instance ID.
This patch allows the operator to specify an "instance ID" for
any particular indielinks intance. This UUID will be used to
disambiguate log messages from other instances in the same cluster.

Since this is implemented in terms of the tracing `#[instrument]`
macro, it's *opt-in*: if you want your log messages to contain
the instance ID, you need to instrument the function at the top
of your call stack.