Sign outgoing ActivityPub requests.
It turns out that Mastodon, when run in "secure mode" (as
indieweb.social apparently is) requires even GET requests to be
signed.

This rather large commit (an amalgam of several smaller ones,
each a "work in progress" commit) does the following:

Introduce "instance state":

This (rather large) patch introduces the idea of "instance state"
to the `schema_migrations` table: a bit of immutable state created at
install time and read by indielinks on startup. As of now, it only
contains the keypair to be used by the instance actor.

This patch does the following:

    - creates `InstanceStateV0`
    - provisions an instance on install
    - loads & verifies it on service startup, adding it to the
      indielinks state object

While at it, I also took care of a few other things I've meant
to do for a long time "while I was there":

    - factor the `Indielinks` struct out of http.rs
    - break-up indielinks-shared/src/lib.rs

Add the instance actor:

This commit introduces the /actor endpoint, which will return
the "instance actor"; an actor representing this indielinks
instance.

Sign all outgoing requests:

This patch signs all outgoing requests, either with the obvious
user, or with the instance actor.