Add a generation to all my cache values.
This patch adds a generation number to each value stored in a
cache. When writing to a given cache, the caller must supply that
number as received on read. If, on write, the number doesn't
match, the cache knows that the caller's value is stale & rejects
the write.

This is another "work in progress" commit. It tests out, but
the code is again littered with TODO comments and the new
logic is (largely) untested. It should be amended and or squashed.